# [SPARK-6195] [SQL] Adds in-memory column type for fixed-precision decimals