--- a/sql/core/src/test/scala/org/apache/spark/sql/columnar/InMemoryColumnarQuerySuite.scala
+++ b/sql/core/src/test/scala/org/apache/spark/sql/columnar/InMemoryColumnarQuerySuite.scala
@@ -17,11 +17,11 @@
 
 package org.apache.spark.sql.columnar
 
-import org.apache.spark.sql.functions._
 import org.apache.spark.sql.TestData._
 import org.apache.spark.sql.catalyst.expressions.Row
 import org.apache.spark.sql.test.TestSQLContext._
 import org.apache.spark.sql.test.TestSQLContext.implicits._
+import org.apache.spark.sql.types.Decimal
 import org.apache.spark.sql.{QueryTest, TestData}
 import org.apache.spark.storage.StorageLevel.MEMORY_ONLY
 
@@ -116,5 +116,17 @@ class InMemoryColumnarQuerySuite extends QueryTest {
     // Shouldn't throw
     complexData.count()
     complexData.unpersist()
+  }
+
+  test("decimal type") {
+    (1 to 10)
+      .map(i => Tuple1(Decimal(i, 20, 10)))
+      .toDF("dec")
+      .cache()
+      .registerTempTable("test_fixed_decimal")
+
+    checkAnswer(
+      sql("SELECT * FROM test_fixed_decimal"),
+      (1 to 10).map(i => Row(Decimal(i, 20, 10).toJavaBigDecimal)))
   }
 }