@liancheng 2015-03-14
Added micro benchmark result. Notice that explicit casting is required because `ScalaReflection` can't capture precision information. This doesn't affect use cases where schema are specified explicitly (Hive tables, parquet tables, etc.).