@liancheng 2015-03-14
/cc @yhuai, this should be helpful for the TPC-DS benchmark. Gonna merge this once Jenkins nods.